Start code generation, further details below
For that, some things were added:
* cc_codegen will hold the functions that generate code
* cc_list is the linked list of instructions/labels

Currently, we generate code for bool, int and float literals and the addition
operation (just for testing).

Also, added a script that runs the cmake with the correct arguments if you give
it the target stage, i.e. ./cmake_helper 5 builds for the 5th stage.